2022-1_psiperfil/pyproject.toml
[tool.poetry]
name = "psiperfil"
version = "0.1.0"
description = ""
authors = ["Felipe Braga"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
pytest = "^7.1.2"
whitenoise = "^6.2.0"
psycopg2-binary = "^2.9.3"
django-heroku = "^0.3.1"
drf-api-logger = "^1.1.10"
gunicorn = "^20.1.0"
Django = "^4.0.5"
django-reset-migrations = "^0.4.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pylint = "^2.14.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = 'scripts.server:start'
makemigrations = 'scripts.server:makemigrations'
migrate = 'scripts.server:migrate'
